pyproject.toml
[tool.poetry]
name = "raylab"
version = "0.16.0"
description = "Reinforcement learning algorithms in RLlib and PyTorch."
authors = ["Ângelo Gregório Lovatto <angelolovatto@gmail.com>"]
license = "MIT"
readme = "README.rst"
repository = "https://github.com/angelolovatto/raylab"
keywords = ["reinforcement-learning", "rllib", "deep-learning", "model-based-rl", "pytorch"]
classifiers = [
"Development Status :: 2 - Pre-Alpha",
"Intended Audience :: Developers",
"Natural Language :: English"
]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
ray = {extras = ["rllib", "tune"], version = "<1.6"}
torch = "^1.5.1"
streamlit = ">=0.62,<0.86"
cachetools = {version = "^4.1.0", python = "^3.7"}
# Have to restrict bokeh version due to issue with streamlit
# https://github.com/streamlit/streamlit/issues/2156#issuecomment-815311729
bokeh = "<2.3.4"
dataclasses-json = "^0.5.1"
opencv-python = "^4.2.0"
poetry-version = "^0.1.5"
pytorch-lightning = ">=1.3.8,<1.5.0"
dm-tree = "^0.1.5"
tabulate = "^0.8.7"
opencv-contrib-python = "^4.4.0"
sklearn = "^0.0"
nnrl = "^0.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pylint = "^2.8.3"
watchdog = "^2.1.2"
black = "^21.6b0"
tox = "^3.24.1"
sphinx = "^4.0.2"
pytest = "^6.2.3"
gym-cartpole-swingup = "^0.1.3"
pre-commit = "^2.13.0"
mypy = "^0.910"
coverage = "^5.5"
ipython = "^7.25.0"
pytest-mock = "^3.6.1"
pytest-sugar = "^0.9.4"
auto-changelog = "^0.5.3"
isort = "^5.9.3"
pytest-deadfixtures = "^2.2.1"

[tool.poetry.scripts]
raylab = "raylab.cli:raylab"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:ray:",
    "ignore::DeprecationWarning:tree:",
    "ignore::DeprecationWarning:tensorflow:",
    "ignore::DeprecationWarning:google:",
    "ignore::DeprecationWarning:flatbuffers:",
    "ignore::DeprecationWarning:wandb:",
    "ignore:pandas.util.testing is deprecated:FutureWarning",
    "ignore::DeprecationWarning:tensorboard:",
    "ignore::DeprecationWarning:pytorch_lightning",
    "ignore::FutureWarning:ray.autoscaler"
]
